Added more styling
