+++ src/App.js
@@ -1,23 +1,14 @@
-import React, { useState, useEffect, useReducer } from "react";
+import React, { useState, useEffect } from "react";
 import Navbar from "./components/Navbar/Navbar";
 import Visualizer from "./components/Visualizer/Visualizer";
 import Settings from "./components/Settings/Settings";
 import useTheme from "./hooks/useTheme";
-import useSpeed from "./hooks/useSpeed";
-import useSorting from "./hooks/useSorting";
 import useArrayGenerator from "./hooks/useArrayGenerator";
-import { changeStructure, INITIAL_STATE } from "./reducers/navReducer";
 
 export default function App() {
-  const [nav, setNav] = useState(false);
+  const [speed, setSpeed] = useState(11);
   const [theme, setTheme] = useTheme();
-  const [speed, setSpeed] = useSpeed();
   const { array, size, generateArray, updateSize } = useArrayGenerator(11);
-  const [state, dispatch] = useReducer(changeStructure, INITIAL_STATE);
-  const { handleSort } = useSorting(state.algo);
-
-  const handleForward = (key) => dispatch({ type: "FORWARD", key });
-  const handleBack = () => dispatch({ type: "BACK" });
 
   const handleSizeChange = (event) => updateSize(parseInt(event.target.value));
   const handleSpeedChange = (event) => setSpeed(parseInt(event.target.value));
@@ -28,19 +19,10 @@ export default function App() {
 
   return (
     <div className={`container ${theme}`}>
-      <Navbar
-        data={state}
-        forward={handleForward}
-        backward={handleBack}
-        theme={theme}
-        setTheme={setTheme}
-        nav={nav}
-        setNav={setNav}
-      />
+      <Navbar theme={theme} setTheme={setTheme} />
       <Visualizer array={array} theme={theme} />
       <Settings
         start={generateArray}
-        sort={() => handleSort(array)}
         size={size}
         speed={speed}
         changeSize={handleSizeChange}
