+++ src/components/Navbar/Navbar.css
@@ -7,59 +7,10 @@
   transition: 0.5s;
 }
 
-.nav-container {
-  position: relative;
-}
-
-.dropdown {
-  display: none;
-  position: absolute;
-  right: 0;
-  top: 100%; /* This ensures the menu appears right below the toggle button */
-  background-color: white; /* Or any color you prefer */
-  border: 1px solid #ccc; /* Just an example for styling */
-  border-radius: 4px;
-  width: 200px; /* Adjust as necessary */
-  list-style: none;
-  padding: 0;
-}
-
-.dropdown.show {
-  display: block;
-}
-
-.logo {
-  width: 160px;
-  cursor: pointer;
-}
-
-ul {
-  flex: 1;
-  list-style: none;
-  text-align: center;
-}
-
-ul li {
-  display: inline-block;
-  margin: 10px 20px;
-  font-size: 18px;
-  cursor: pointer;
-}
-
 .toggle-icon {
-  width: 50px;
+  width: 4rem;
   cursor: pointer;
   margin-left: 10px;
-}
-
-.icon {
-  background: none;
-  cursor: pointer;
-  border: none;
-  font-size: 2em;
-  z-index: 1;
-  margin-right: 10px;
-  color: #000;
 }
 
 .search-box {
@@ -89,6 +40,57 @@ ul li {
   cursor: pointer;
 }
 
+.nav {
+  position: relative;
+}
+
+.icon {
+  background: none;
+  cursor: pointer;
+  border: none;
+  font-size: 2em;
+  z-index: 1;
+  margin-right: 10px;
+  color: #000;
+  padding: 5px 20px;
+}
+
+.dropdown {
+  display: none;
+  position: absolute;
+  text-align: center;
+  right: 0;
+  background-color: #242424;
+  border: 1px solid #ccc;
+  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
+  border-radius: 15px;
+  width: 200px;
+  list-style: none;
+  margin-top: 10px;
+}
+
+.dropdown.show {
+  display: block;
+}
+
+.item {
+  padding: 5px;
+}
+
+.item button {
+  background: none;
+  border: none;
+  cursor: pointer;
+  color: #fff;
+}
+
+.item:hover {
+  box-shadow: 2px 2px 8px 2px rgba(184, 177, 177, 0.2);
+  border-radius: 50px;
+}
+
+/*  Dark Mode  */
+
 .dark .navbar {
   background: #000;
   color: #fff;
@@ -101,25 +103,25 @@ ul li {
 .dark .search-box input {
   color: #222;
 }
+
 .dark .search-box input::placeholder {
-  color: #222;
+  color: #000;
 }
 
 .dark .icon {
   color: #fff;
 }
 
-/*  Navbar  */
-
-.dropdown {
-  box-shadow: 0 8px 16px 0 rgba(0, 0, 0, 0.2);
-  padding: 1rem;
-  right: 0;
-  position: absolute;
+.dark .dropdown {
+  background-color: #fff;
+  border: 1px solid #ccc;
 }
 
-.item button {
-  background: none;
-  border: none;
-  cursor: pointer;
+.dark .item button {
+  color: #000;
+}
+
+.dark .item:hover {
+  box-shadow: 8px 8px 8px 8px rgba(0, 0, 0, 0.2);
+  color: #000;
 }
