+++ src/components/Settings/Settings.jsx
@@ -2,40 +2,41 @@ import "./Settings.css";
 
 export default function Settings({
   start,
-  sort,
   size,
   speed,
   changeSize,
   changeSpeed,
   theme,
 }) {
+  const renderSlider = (name, value, min, max, onChange) => {
+    return (
+      <div>
+        <p className="slider">
+          <span>{name}</span>
+          <span>{value}</span>
+        </p>
+        <input
+          type="range"
+          className={`${name}-slider`}
+          min={min}
+          max={max}
+          value={value}
+          onChange={onChange}
+        />
+      </div>
+    );
+  };
+
   return (
     <div id="settings">
       <button className={`settings-btn ${theme}`} onClick={start}>
-        Start
+        Shuffle
       </button>
 
-      <button className={`settings-btn ${theme}`} onClick={sort}>
-        Sort
-      </button>
+      <button className={`settings-btn ${theme}`}>Sort</button>
 
-      <input
-        type="range"
-        className="slider"
-        min="11"
-        max="111"
-        value={size}
-        onChange={changeSize}
-      />
-
-      <input
-        type="range"
-        className="slider"
-        min="-111"
-        max="1111"
-        value={speed}
-        onChange={changeSpeed}
-      />
+      {renderSlider("size", size, 11, 111, changeSize)}
+      {renderSlider("speed", speed, -111, 1111, changeSpeed)}
     </div>
   );
 }
