--- src/library/fake.js
@@ -1,16 +1,0 @@
-const fake = {
-  Bubble: () => {
-    console.log("Bubble");
-  },
-  Insertion: () => {
-    console.log("Insertion");
-  },
-  Merge: () => {
-    console.log("Merge");
-  },
-  Quick: () => {
-    console.log("Quick");
-  },
-};
-
-export default fake;
